Added services and controllers. Code largely untested and not working properly.

+++ src/main/java/se/lexicon/laserbrain1613/booklender/service/LoanService.java
@@ -20,5 +20,5 @@ public interface LoanService {
 
     LoanDto update(LoanDto loanDto);
 
-    boolean delete(int loanId);
+    boolean delete(long loanId);
 }
